examples/cmdb/backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
SQLAlchemy==2.0.48
PyYAML==6.0.3
pytest==9.0.2
httpx==0.28.1
safrs==3.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.22
